+++ R/plotStackProfile.R
@@ -54,7 +54,6 @@
 #' @return Display a plot.
 #'
 #' @importFrom S4Vectors mcols
-#' @importFrom viridisLite viridis
 #' @importFrom BiocGenerics image
 #' @importFrom graphics axis
 #' @importFrom graphics mtext
@@ -73,7 +72,7 @@ plotStackProfile <- function(
     x_labels = c("Before", "Anchor", "After"),
     title = "",
     zlim = NULL,
-    palette = function(n) viridisLite::viridis(n, direction = -1),
+    palette = function(n) hcl.colors(n, rev = TRUE),
     target_height = 650,
     summary_func = function(x) mean(x, na.rm = TRUE),
     n_core = 1,